requirements.txt
asn1crypto==0.24.0
astroid==2.2.0
attrs==18.2.0
Automat==0.7.0
cachetools==3.1.0
certifi==2018.11.29
cffi==1.12.2
chardet==3.0.4
constantly==15.1.0
cryptography==2.6.1
cssselect==1.0.3
google-api-core==1.8.0
google-auth==1.6.3
google-cloud-vision==0.36.0
googleapis-common-protos==1.5.8
grpcio==1.19.0
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
isort==4.3.9
lazy-object-proxy==1.3.1
lxml==4.3.1
mccabe==0.6.1
parsel==1.5.1
protobuf==3.6.1
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==1.9.0
pylint==2.3.0
pyOpenSSL==19.0.0
pytz==2018.9
queuelib==1.5.0
requests==2.21.0
rsa==4.0
Scrapy==1.6.0
selenium==3.141.0
service-identity==18.1.0
six==1.12.0
Twisted==18.9.0
typed-ast==1.3.1
urllib3==1.24.1
w3lib==1.20.0
wrapt==1.11.1
zope.interface==4.6.0
